processor/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
PyYAML==6.0.2
qdrant-client==1.11.3
sentence-transformers==3.2.1
numpy==1.26.4
pypdf==5.0.1
python-pptx==1.0.2
watchdog==5.0.3
requests==2.32.3
python-multipart==0.0.10
faster-whisper==1.0.3
Pillow==10.4.0
opencv-python-headless==4.10.0.84
paddleocr==2.9.1
scikit-image==0.24.0
python-dotenv==1.0.1
pyannote.audio==3.3.2
pytest==8.3.3
